fix(simulator): preserve paged decode index dtype
Cast paged decode indices before assigning int32 free pages, and cover a page-boundary decode with output_length greater than one. Keep cache-tier assertions in the cache-specific suite and support both current and legacy HiCache host-pool interfaces.

Move the simulator-aware serving benchmark to benchmark/simulator and keep its source path wired into the extra CPU CI gate.

+++ tools/sglang-simulator/src/sglang_simulator/simulation/sglang/mem_cache_allocator.py
@@ -66,14 +66,14 @@ def _alloc_decode_cpu(
     if num_new_pages > len(self.free_pages):
         return None
 
-    out_indices = last_loc + 1
+    out_indices = (last_loc + 1).to(dtype=self.free_pages.dtype)
     need_new_page = seq_lens % self.page_size == 1
     if num_new_pages:
         out_indices = out_indices.clone()
         out_indices[need_new_page] = self.free_pages[:num_new_pages] * self.page_size
 
     self.free_pages = self.free_pages[num_new_pages:]
-    return out_indices.to(self.free_pages.dtype)
+    return out_indices
 
 
 def alloc_extend_cpu(*args, **kwargs):
